add.js
Critical change to fix !add critically failing should noblox not be logged in and able to promote the user. There is now a catch on the promotion process to ensure that the app does not fail to give experience to users if they are unable to be promoted at the moment. Will be updated eventually to be more secure.